requirements.txt
fastapi==0.127.0
groq==1.0.0
python-dotenv==1.0.1
uvicorn==0.22.0
python-multipart==0.0.21
httpx==0.28.1
playwright==1.58.0